+++ tests/RecordTests/UnitTest/sample_recordings.cpp
@@ -63,7 +63,7 @@ void SampleRecordings::SetUp()
         uint64_t imu_timestamp = 1150;
         uint32_t timestamp_delta = 1000000 / k4a_convert_fps_to_uint(record_config_full.camera_fps);
         k4a_capture_t capture = NULL;
-        for (int i = 0; i < test_frame_count; i++)
+        for (size_t i = 0; i < test_frame_count; i++)
         {
             capture = create_test_capture(timestamps,
                                           record_config_full.color_format,
@@ -106,7 +106,7 @@ void SampleRecordings::SetUp()
                                    (uint64_t)record_config_delay.depth_delay_off_color_usec };
         uint32_t timestamp_delta = 1000000 / k4a_convert_fps_to_uint(record_config_delay.camera_fps);
         k4a_capture_t capture = NULL;
-        for (int i = 0; i < test_frame_count; i++)
+        for (size_t i = 0; i < test_frame_count; i++)
         {
             capture = create_test_capture(timestamps,
                                           record_config_delay.color_format,
@@ -166,7 +166,7 @@ void SampleRecordings::SetUp()
 
         uint64_t timestamps[3] = { 1000000, 1001000, 1001000 }; // Start recording at 1s
         uint32_t timestamp_delta = 1000000 / k4a_convert_fps_to_uint(record_config_full.camera_fps);
-        for (int i = 0; i < test_frame_count; i++)
+        for (size_t i = 0; i < test_frame_count; i++)
         {
             // Create a known pattern of dropped / missing frames that can be tested against
             // The pattern is repeated every 4 captures until the end of the file.
